% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koina-class.R
\docType{class}
\name{Koina-class}
\alias{Koina-class}
\alias{Koina}
\title{koina class}
\value{
an instance of the Koina class
}
\description{
koina class
}
\section{Fields}{

\describe{
\item{\code{model_inputs}}{list.}

\item{\code{model_outputs}}{list.}

\item{\code{batch_size}}{numeric.}

\item{\code{response_dict}}{list.}

\item{\code{model_name}}{character.}

\item{\code{url}}{character.}

\item{\code{ssl}}{logical.}

\item{\code{disable_progress_bar}}{logical.}

\item{\code{client}}{ANY.}

\item{\code{type_convert}}{list.}
}}


\examples{
library(koinar)
## Example instantiation of the Koina class
koina_instance <- Koina$new(
  model_name = "Prosit_2019_intensity",
  server_url = "koina.wilhelmlab.org:443",
  ssl = TRUE
)

## Example input data for predict_batch
## TODO(cp): detect dim e.g., use length of peptide 
input <- list(
  peptide_sequences = array(c("LGGNEQVTR", "GAGSSEPVTGLDAK"),
    dim = c(2, 1)),
  collision_energies = array(c(25, 25), dim = c(2, 1)),
  precursor_charges = array(c(1, 2), dim = c(2, 1))
)

prediction_results <- koina_instance$predict(input)
}
\seealso{
\url{https://koina.wilhelmlab.org/docs}
}
\author{
Ludwig Lautenbacher, 2024
}
