% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koina-class.R
\docType{class}
\name{Koina-class}
\alias{Koina-class}
\alias{Koina}
\title{koina class}
\value{
an instance of the Koina class
}
\description{
koina class
}
\section{Fields}{

\describe{
\item{\code{model_inputs}}{list TODO.}

\item{\code{model_outputs}}{list TODO.}

\item{\code{batch_size}}{numeric TODO.}

\item{\code{response_dict}}{list TODO.}

\item{\code{model_name}}{character, e.g., \code{"AlphaPept_ms2_generic"} or \code{"Prosit_2019_intensity"}.}

\item{\code{url}}{url, default is set to \code{"koina.wilhelmlab.org:443"}.}

\item{\code{ssl}}{logical.}

\item{\code{disable_progress_bar}}{logical.}

\item{\code{client}}{ANY.}

\item{\code{type_convert}}{list TODO.}
}}

\section{Methods}{

\describe{
\item{\code{predict(input_data, pred_as_df = TRUE, min_intensity = 1e-05)}}{predict using the defined model TODO}
}}

\examples{
library(koinar)
prosit2019 <- koinar::Koina$new(
  model_name = "Prosit_2019_intensity",
  server_url = "koina.wilhelmlab.org:443"
)

input <- data.frame(
  peptide_sequences = c("LGGNEQVTR", "GAGSSEPVTGLDAK"),
  collision_energies = c(25, 25),
  precursor_charges = c(1, 2)
)

# Fetch the predictions by calling `$predict` of the model you want to use
prediction_results <- prosit2019$predict(input)
}
\seealso{
\url{https://koina.wilhelmlab.org/docs}
}
\author{
Ludwig Lautenbacher, 2024
}
