% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koina-class.R
\docType{class}
\name{Koina-class}
\alias{Koina-class}
\alias{Koina}
\title{koina class}
\arguments{
\item{list_of_list_of_arrays}{A list of lists, where each inner list represents a "batch"
containing named arrays (either vectors or matrices) to be aggregated.}
}
\value{
an instance of the Koina class

A list of aggregated arrays, where each named element is either a matrix 
        (for originally matrix-type elements) or a 1-column matrix (for vector-type elements),
        aggregating corresponding elements across all batches.
}
\description{
koina class
}
\section{Fields}{

\describe{
\item{\code{model_inputs}}{list.}

\item{\code{model_outputs}}{list.}

\item{\code{batch_size}}{numeric.}

\item{\code{response_dict}}{list.}

\item{\code{model_name}}{character.}

\item{\code{url}}{character.}

\item{\code{ssl}}{logical.}

\item{\code{disable_progress_bar}}{logical.}

\item{\code{client}}{ANY.}

\item{\code{type_convert}}{list.}
}}


\examples{
library(koinar)
## Example instantiation of the Koina class
koina_instance <- Koina$new(
  model_name = "Prosit_2019_intensity",
  server_url = "koina.wilhelmlab.org:443",
  ssl = TRUE
)

## Example input data for predict_batch
## TODO(cp): detect dim e.g., use length of peptide 
input <- list(
  peptide_sequences = array(c("LGGNEQVTR", "GAGSSEPVTGLDAK"),
    dim = c(2, 1)),
  collision_energies = array(c(25, 25), dim = c(2, 1)),
  precursor_charges = array(c(1, 2), dim = c(2, 1))
)

prediction_results <- koina_instance$predict(input)
Aggregate Batches of Arrays

Takes a list of lists containing arrays and aggregates them by name across all batches. 
For matrices, they are concatenated along rows. Vectors are treated as 1-column matrices.

# Assuming batch1 and batch2 are lists containing named matrices/vectors
aggregate_batches(list(batch1, batch2))

}
\seealso{
\url{https://koina.wilhelmlab.org/docs}
}
\author{
Ludwig Lautenbacher, 2024
}
