% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koina-class.R
\docType{class}
\name{Koina-class}
\alias{Koina-class}
\alias{Koina}
\title{koina class}
\value{
an instance of the Koina class
}
\description{
koina class
}
\section{Fields}{

\describe{
\item{\code{model_inputs}}{list.}

\item{\code{model_outputs}}{list.}

\item{\code{batch_size}}{numeric.}

\item{\code{response_dict}}{list.}

\item{\code{model_name}}{character.}

\item{\code{url}}{character.}

\item{\code{ssl}}{logical.}

\item{\code{disable_progress_bar}}{logical.}

\item{\code{client}}{ANY.}

\item{\code{type_convert}}{list.}
}}


\examples{
library(koinar)
prosit2019 <- koinar::Koina$new(
  model_name = "Prosit_2019_intensity",
  server_url = "koina.wilhelmlab.org:443"
)

input <- data.frame(
  peptide_sequences = c("LGGNEQVTR", "GAGSSEPVTGLDAK"),
  collision_energies = c(25, 25),
  precursor_charges = c(1, 2)
)

# Fetch the predictions by calling `$predict` of the model you want to use
prediction_results <- prosit2019$predict(input)
}
\seealso{
\url{https://koina.wilhelmlab.org/docs}
}
\author{
Ludwig Lautenbacher, 2024
}
